+++ R/crssi_create_dnf_files.R
@@ -1,38 +1,41 @@
 #' Create CRSS Natural Flow Input Files 
 #' 
-#' Creates the CRSS natural flow input files used by the Colorado River Simulation
-#' System (CRSS)
+#' Creates the CRSS natural flow input files used by the Colorado River 
+#' Simulation System (CRSS)
 #' 
 #' `crssi_create_dnf_files()` and `createCRSSDNFInputFiles()` create individual 
 #' trace files for the observed natural flow on the Colorado River Basin using the 
 #' \href{http://onlinelibrary.wiley.com/doi/10.1111/j.1752-1688.1997.tb03557.x/abstract}{Index Sequential Method}. 
 #' Trace files are formated and saved in a format expected
 #' by the Colorado River Simulation System (CRSS). Data is read from the natural
-#' flow workbook available at \url{http://www.usbr.gov/lc/region/g4000/NaturalFlow/current.html} or
-#' from the `CoRiverNF` data package. It is much faster to use the data package
-#' rather than the Excel workbook. The data package is created from the Excel
-#' file, so the files created are identical for each of the two data sets.
+#' flow workbook available at \url{http://www.usbr.gov/lc/region/g4000/NaturalFlow/current.html} 
+#' or from the [CoRiverNF](https://github.com/BoulderCodeHub/CoRiverNF) data 
+#' package. It is much faster to use the data package rather than the Excel 
+#' workbook. The data package is created from the Excel file, so the files 
+#' created are identical for each of the two data sets.
 #' 
-#' In addition to creating the natural flow input files, data for two informative 
-#' slots are created. These slots include the trace number and the supply scenario
-#' number. The trace numbers are intuitive, i.e., they are integers from 1 to N
-#' where N is the trace number. This information is saved in the slot name that
-#' is set by the "crssio.traceNumberSlot" option. The scenario number provides
-#' the user with a numeric representation of which supply scenario is used. The
-#' observed historical natural flows (used by this function) are supply scenario 1. 
-#' For this supply scenario, the decimals of the supply scenario number represent
-#' the start and end year that the ISM method are applied to. For example, if 
-#' you set \code{recordToUse} to \code{c('1988-1','2012-12')}, the decimal portion
-#' will be 19882012, where the first 4 numbers represent the start year and
-#' the second four numbers represent the end year. The supply scenario slot will 
-#' be set to 1.19882012 in this example. This tells the user of CRSS that the 
-#' supply scenario is the observed historical natural flows with the ISM method
-#' applied to the 1988-2012 data. The supply scenario slot name is set
-#' by the "crssio.supplyScenarioSlot" option.
+#' In addition to creating the natural flow input files, data for two 
+#' informative slots are created. These slots include the trace number and the 
+#' supply scenario number. The trace numbers are intuitive, i.e., they are 
+#' integers from 1 to N where N is the trace number. This information is saved 
+#' in the slot name that is set by the "crssio.traceNumberSlot" option. The 
+#' scenario number provides the user with a numeric representation of which 
+#' supply scenario is used. The observed historical natural flows (used by this 
+#' function) are supply scenario 1. For this supply scenario, the decimals of 
+#' the supply scenario number represent the start and end year that the ISM 
+#' method are applied to. For example, if you set `recordToUse` to 
+#' `c('1988-1','2012-12')`, the decimal portion will be 19882012, where the
+#' first 4 numbers represent the start year and the second four numbers 
+#' represent the end year. The supply scenario slot will be set to 1.19882012 in 
+#' this example. This tells the user of CRSS that the  supply scenario is the 
+#' observed historical natural flows with the ISM method applied to the 
+#' 1988-2012 data. The supply scenario slot name is set by the 
+#' "crssio.supplyScenarioSlot" option.
 #' 
 #' The hydrologyIncrement data that sets the random number generator for
-#' each year and trace is created for each trace folder. The slot name that is created
-#' for the hydrologyIncrement is set by the "crssio.hydroIncrement" option.
+#' each year and trace is created for each trace folder. The slot name that is 
+#' created for the hydrologyIncrement is set by the "crssio.hydroIncrement" 
+#' option.
 #' 
 #' Beginning in CRSS v2.6, input data for the Sacramento year type 
 #' index are necessary. The historical Sacramento year type data (available at
@@ -45,10 +48,11 @@
 #' `overwriteFiles` allows the user to control whether existing files within the
 #' trace folders should be overwritten (default is they are not). 
 #' 
-#' @param iFile Either the string "CoRiverNF", or the relative or absolute path 
-#' to the excel workbook. When "CoRiverNF" is used, the data from the `CoRiverNF`
-#' data package is used. Otherwise, it should be a valid path to the natural 
-#' flow Excel workbook. 
+#' @param iFile Either the string `"CoRiverNF"``, or the relative or absolute 
+#'   path to the excel workbook. When "CoRiverNF" is used, the data from the 
+#'   [CoRiverNF](https://github.com/BoulderCodeHub/CoRiverNF) data package is 
+#'   used. Otherwise, it should be a valid path to the natural flow Excel 
+#'   workbook. 
 #' @param oFolder Path to the top level directory where the trace folders and
 #'   input files will be created. This folder should exist before using this
 #'   function.