+++ tests/testthat/test_crsso_sys_cond_table.R
@@ -51,7 +51,9 @@ test_that("warnings and errors are as expected", {
     fixed = TRUE
   )
   expect_error(
-    sysData %>% filter(Variable != "mer748", Variable != "eq") %>% crsso_get_sys_cond_table(yrs),
+    sysData %>% 
+      filter(Variable != "mer748", Variable != "eq") %>% 
+      crsso_get_sys_cond_table(yrs),
     "The following variables are not found in the data frame passed to crsso_get_sys_cond_table():\nmer748, eq",
     fixed = TRUE
   )
@@ -70,35 +72,35 @@ test_that("warnings and errors are as expected", {
   )
 })
 
-expVals <- matrix(c(
-  rep(0, 5),
-  rep(0, 5),
-  rep(0, 5),
-  c(100, 100, 25, 25, 50),
-  c(100, 100, 0, 0, 50),
-  c(0, 0, 25, 25, 0), 
-  rep(0, 5),
-  c(0, 0, 75, 75, 25),
-  rep(0, 5),
-  c(0, 0, 75, 75, 25),
-  rep(0, 4), 25,
-  rep(0, 5), 
-  rep(0, 5),
-  rep(0, 4), 25,
-  c(0, 50, 75, 75, 100),
-  c(0, 50, 75, 25, 50),
-  c(0, 0, 0, 50, 50),
-  rep(0, 5),
-  rep(0, 5),
-  rep(0, 5),
-  c(100, 50, 25, 25, 0)),
+expVals <- matrix(
+  c(
+    rep(0, 5), # EQ
+    rep(0, 5), # EQ > 8.23
+    rep(0, 5), # EQ = 8.23
+    c(100, 100, 25, 25, 50), # UEB
+    c(100, 100, 0, 0, 50), # UEB > 8.23
+    c(0, 0, 25, 25, 0),  # UEB = 8.23
+    rep(0, 5), # UEB < 8.23
+    c(0, 0, 75, 75, 25), # Mid
+    rep(0, 5), # Mid = 8.23
+    c(0, 0, 75, 75, 25), # Mid = 7.48
+    rep(0, 4), 25, # LEB
+    rep(0, 5), # LEB > 8.23
+    rep(0, 5), # LEB = 8.23
+    rep(0, 4), 25, # LEB < 8.23
+    rep(0, 5), # surplus
+    rep(0, 5), # FC Surplus
+    c(100, 50, 25, 25, 0), # Normal
+    c(0, 50, 75, 75, 100), # short any
+    c(0, 50, 75, 25, 50), # short 1
+    c(0, 0, 0, 50, 50), # short 2
+    rep(0, 5) # short 3
+  ),
   ncol = 5, 
   byrow = T
 )
 
-
-
-# thinkt hat checking against the orig values, and then selecting a few values
+# think that checking against the orig values, and then selecting a few values
 # to recompute should test it enough
 # check UEB Total, Shortage 1, MER 7.48 and normal year
 r1 <- apply(
@@ -119,11 +121,13 @@ r1 <- apply(
   1,
   mean
 ) * 100
+
 r2 <- apply(
   rdf_get_slot(RWDataPlyr::sysRdf, "SummaryOutputData.LBShortageStep1"),
   1,
   mean
 ) * 100
+
 r3 <- apply(
   rdf_get_slot(
     RWDataPlyr::sysRdf, 
@@ -132,49 +136,63 @@ r3 <- apply(
   1,
   mean
 ) * 100
+
 r4 <- apply(
   rdf_get_slot(RWDataPlyr::sysRdf, "SummaryOutputData.LBNormalCondition"),
   1,
   mean
 ) * 100
+
 test_that("computations of chances are correct", {
- expect_equivalent(expVals, sysCondTable$fullTable)
- expect_equivalent(sysCondTable$fullTable[4,], r1)
- expect_equivalent(sysCondTable$fullTable[16,], r2)
- expect_equivalent(sysCondTable$fullTable[10,], r3)
- expect_equivalent(sysCondTable$fullTable[21,], r4)
+  expect_equivalent(expVals, sysCondTable$fullTable)
+  # UEB Total
+  expect_equivalent(sysCondTable$fullTable[4,], r1)
+  # Shortage 1
+  expect_equivalent(sysCondTable$fullTable[19,], r2)
+  # MER 7.48
+  expect_equivalent(sysCondTable$fullTable[10,], r3)
+  # normal year
+  expect_equivalent(sysCondTable$fullTable[17,], r4)
 })
 
 test_that("rows sum together correctly", {
+  # EQ
   expect_equal(
     sysCondTable$fullTable[1,], 
     apply(sysCondTable$fullTable[2:3,], 2, sum)
   )
+  # UEB
   expect_equal(
     sysCondTable$fullTable[4,], 
     apply(sysCondTable$fullTable[5:7,], 2, sum)
   )
+  # MER
   expect_equal(
     sysCondTable$fullTable[8,], 
     apply(sysCondTable$fullTable[9:10,], 2, sum)
   )
+  # LEB
   expect_equal(
     sysCondTable$fullTable[11,], 
     apply(sysCondTable$fullTable[12:14,], 2, sum)
   )
+  # Shortage
   expect_equal(
-    sysCondTable$fullTable[15,], 
-    apply(sysCondTable$fullTable[16:18,], 2, sum)
+    sysCondTable$fullTable[18,], 
+    apply(sysCondTable$fullTable[19:21,], 2, sum)
   )
+  # All Powell Tiers should sum to 100
   expect_equivalent(
     rep(100, 5), 
-    apply(sysCondTable$fullTable[c(1,4,8,11),], 2, sum)
+    apply(sysCondTable$fullTable[c(1, 4, 8, 11),], 2, sum)
   )
+  # All Lower Basin Tiers should sum to 100
   expect_equivalent(
     rep(100, 5), 
-    apply(sysCondTable$fullTable[c(15,19,21),], 2, sum)
+    apply(sysCondTable$fullTable[c(15, 17, 18),], 2, sum)
   )
-  expect_true(all(sysCondTable$fullTable[20,] <= sysCondTable$fullTable[19,]))
+  # FC surplus should always be <= the total surplus
+  expect_true(all(sysCondTable$fullTable[16,] <= sysCondTable$fullTable[15,]))
 })
 
 # compare both versions of the function ---------------------